remove redux as it's not needed
